# ZIR-153: Introduce directive to manually alias layouts