--- a/zirgen/Conversions/Typing/ZhlComponent.cpp
+++ b/zirgen/Conversions/Typing/ZhlComponent.cpp
@@ -927,6 +927,40 @@ Value LoweringImpl::asLayout(Value value) {
       .Case<BackOp>([&](BackOp op) {
         return asLayout(op.getTarget());
       })
+      .Case<ArrayOp>([&](ArrayOp op) {
+        SmallVector<Value> layouts;
+        for (Value element : op.getElements())
+          layouts.push_back(asLayout(element));
+        return builder.create<ZStruct::LayoutArrayOp>(op.getLoc(), layouts);
+      })
+      .Case<MapOp>([&](MapOp op) {
+        Value array = asValue(op.getArray());
+        Type elemType = cast<ArrayType>(array.getType()).getElement();
+        Region mapBody(regionAnchor);
+        Type layoutType;
+        {
+          OpBuilder::InsertionGuard insertionGuard(builder);
+          Block* mapBodyBlock = builder.createBlock(&mapBody);
+          auto mapArg = op.getFunction().getArgument(0);
+          valueMapping[mapArg] = mapBodyBlock->addArgument(elemType, mapArg.getLoc());
+          auto super = cast<SuperOp>(op.getFunction().back().getTerminator());
+          auto layout = asLayout(super.getValue());
+          layoutType = layout.getType();
+          builder.create<ZStruct::YieldOp>(super->getLoc(), layout);
+        }
+        size_t size = cast<ArrayType>(array.getType()).getSize();
+        Type layoutArrayType = LayoutArrayType::get(ctx, layoutType, size);
+        auto map = builder.create<ZStruct::MapOp>(op->getLoc(), layoutArrayType, array, Value());
+        map.getBody().takeBody(mapBody);
+        return map;
+      })
+      .Case<BlockOp>([&](BlockOp op) {
+        // If a block has layout for its members, it will be associated in the
+        // layoutMapping during lowering. At this point that must not be the
+        // case so look for the super's layout coming from outside the block.
+        auto super = cast<SuperOp>(op.getInner().back().getTerminator());
+        return asLayout(super.getValue());
+      })
       .Default([&](Operation* op) {
         llvm::outs() << "unhandled op: " << *op << "\n";
         return nullptr;
@@ -943,7 +977,7 @@ void LoweringImpl::gen(DirectiveOp directive, ComponentBuilder& cb) {
     }
     Value left = asLayout(directive.getArgs()[0]);
     Value right = asLayout(directive.getArgs()[1]);
-    assert(left);
+    assert(left && right);
     Type type = Zhlt::getLeastCommonSuper({left.getType(), right.getType()}, /*isLayout=*/1);
     left = coerceTo(left, type);
     right = coerceTo(right, type);