--- a/zirgen/dsl/test/repro/back_arg_segfault.zir
+++ b/zirgen/dsl/test/repro/back_arg_segfault.zir
@@ -1,0 +1,10 @@
+// RUN: zirgen %s --emit=zhlt
+
+component Foo(rt1: Reg) {
+    Reg(5)
+}
+
+component Top() {
+    x : Reg;
+    x := Foo(x@1);
+}