@jacobdweightman 2024-09-05
Added some performance enhancements that go a long way in counteracting the performance hit of adding this feature on the current draft of the Keccak accelerator. These are the numbers I ended up with on my machine for the "small test" (2897 cycles) in the interpreter:

Before: 106 s
After (no performance work): 402 s
After (with performance work):  209 s

I think there's a good deal more "easy" performance work, which I'd like to poke at next. I hope it might be possible to get another 2x from further parallelizing the compilation pipeline, but I'd like to move towards merging this soon.