@jacobdweightman 2024-09-05
Since the bulk of the time was spent canonicalizing for the interpreter, I just tried turning off that step which presumably results in us interpreting a lot more instructions, but skips that apparently expensive pass. On my machine, I just ran the short Keccak test in the interpreter in 74 seconds, which is about ~25% faster than before.