@jacobdweightman 2024-09-06
Pushed one more commit that fixes a regression in the v2 circuit; talked with Nils about an alternative approach, but we're going to merge now anyway and I'll follow up.